==> 2. Preparing the data for classification.ipynb <==
To make the classifier as accurate as possible it is neccesery to train it with at much data as possible.
Unfortunately classifying all the files on an smb share is too much effort.

In this chapter we will classify a certain part of the dataset which will be used to learn from the rest of the data available.


import pandas as pd

First lets import the file created in the previous chapter.
This dataframe will be altered to a new format which will be used in the process of writing the classifier.

df = pd.read_pickle("data/bro_logs_filtered.pkl")

files_df = pd.Series(df["name"].unique())

files_df = files_df[files_df.str[-5:].str.contains("\.")]

files_df = files_df.to_frame("name")



files_df["label"] = None

print files_df.shape

Since not all companies are equal, their labels differ as well.
Having more labels will give a more accurate fingerprint on user behavior but having not enough information about a label for the classifier will result in no classification at all.

Try to use labels which are generic enough to classify but specific enough to see changes in behavior such as public, customer info and internal.

files_df.to_csv("data/custom_files_labels.csv")

Now that we have part of the files labeled we can continue with semi supervised learning

files_df_labeled = pd.read_csv("data/custom_files_labels_done.csv", na_values=[])

files_df_labeled["label"].fillna('', inplace=True)

            
print("{} out of {} Files labeled in dataset").format(len(files_df_labeled[files_df_labeled["label"] != '']), len(files_df))
print files_df_labeled["label"][files_df_labeled["label"] != ''].value_counts()
print files_df_labeled.head(25)
            

print files_df_labeled.dtypes


files_df_labeled.to_pickle("data/custom_files_labels_done.pkl")
