==> 4. Meaningfull logging.ipynb <==
To provide meaningfull logging we need to provide the customer with names instead of IP adresses

Doing this will not only speed up the process of finding a user and malicous activities but will also prevent false changes due to IP leases.

This code will add a row to the dataframe for now to show how it works but should perform realtime during live monitoring.

import pandas as pd

smb_cmd = pd.read_csv("data/smb_cmd.csv", sep="\t", na_values="-", parse_dates=[0])
smb_cmd.columns = [ts, uid, id.orig_h, id.orig_p, id.resp_h, id.resp_p, command, sub_command, argument, status, rtt, version, username, tree, tree_service, blob]
kerberos.fillna(0, inplace=True)
print smb_cmd.shape

print smb_cmd[smb_cmd["username"] != 0].shape
print smb_cmd[smb_cmd["username"] != 0].value_counts()

kerberos = pd.read_csv("data/kerberos.csv", sep="\t", na_values="-", parse_dates=[0, 11, 12])
kerberos.columns = ["ts", "uid", "id.orig_h", "id.orig_p", "id.resp_h", "id.resp_p", "request_type", "client", "service", "success", "error_msg", "from", "till", "cipher", "forwardable", "renewable", "client_cert_subject", "client_cert_fuid", "server_cert_subject", "server_cert_fuid"]
kerberos.fillna(0, inplace=True)
print kerberos.shape

def ip_to_name_kerberos(ip):
    try:
        maxloc =  kerberos["ts"][kerberos["id.orig_h"] == ip][kerberos["client"] != 0].argmax()
        name = kerberos.loc[maxloc]["client"]
        return name.split("/")[0].lower()
    except:
        return "unknown"

print ip_to_name("IP address")

bro_logs = pd.read_pickle("data/bro_logs_filtered.pkl")

bro_logs_read = bro_logs[bro_logs["action"] == "SMB::FILE_READ"]
print bro_logs_read.head(10)
bro_logs_read["user"] = bro_logs_read["id.orig_h"].apply(ip_to_name)
print bro_logs_read.head(10)

print bro_logs_read["user"].value_counts()
